refactor: feedcomponent
